% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort_raw.R
\docType{data}
\name{cohort_raw}
\alias{cohort_raw}
\title{A simulated dataset containing the survival data of 100 patients}
\format{
A data frame with 1626 rows and 2 variables:
\describe{
  \item{patientid}{character. Unique patientid}
  \item{sex}{character. Patient sex}
  \item{age}{numeric. Patient's age at index}
  \item{condition}{character. The condition a patient was in.}
  \item{lotstartdate}{date. Date at which first line of therapy starts. used
  as an indexdate}
  \item{lastvisitdate}{date. Last known patient contactdate}
  \item{dateofdeath}{date. Date of death if known.}
}
}
\usage{
cohort_raw
}
\description{
A simulated dataset containing the survival data of 100 patients
}
\keyword{datasets}
