% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_event_status.R, R/derive_functions.R
\name{derive_event_status}
\alias{derive_event_status}
\title{Calculate event_status, a logical column indicating whether or not an event occured.}
\usage{
derive_event_status(
  data = NULL,
  event_date = NULL,
  event_status_rename = "event_status"
)

derive_event_status(
  data = NULL,
  event_date = NULL,
  event_status_rename = "event_status"
)
}
\arguments{
\item{data}{dataframe. A one-row-per-patient dataframe.}

\item{event_date}{character. The name of a column in data indicating event dates.}

\item{event_status_rename}{character. (Optional) Name of the resulting event_status column}
}
\value{
dataframe.

dataframe.
}
\description{
Calculate event_status, a logical column indicating whether or not an event occured.

Calculate event_status, a logical column indicating whether or not an event occured.
}
