% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_follow_up_time.R
\name{derive_follow_up_date}
\alias{derive_follow_up_date}
\title{Calculate follow up date from event_date and censor_date columns}
\usage{
derive_follow_up_date(
  cohort = NULL,
  event_date = NULL,
  censor_date = "censor_date"
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe.}

\item{event_date}{character. The name of a column in cohort indicating event dates.}

\item{censor_date}{character. The name of a date column in cohort indicating censor dates.}
}
\value{
dataframe.
}
\description{
Calculate follow up date from event_date and censor_date columns
}
