% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_follow_up_time.R
\name{derive_follow_up_time}
\alias{derive_follow_up_time}
\title{Calculate follow_up_time, the number of days from index to follow up in days}
\usage{
derive_follow_up_time(
  cohort,
  index_date = "index_date",
  follow_up_date = "follow_up_date",
  follow_up_time_rename = "follow_up_days"
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe.}

\item{index_date}{character. The name of a column in cohort indicating index dates.}

\item{follow_up_date}{character. The name of a column in cohort indicating follow up dates.}

\item{follow_up_time_rename}{character. (Optional) Name of the resulting follow up days column.}
}
\value{
dataframe.
}
\description{
Calculate follow_up_time, the number of days from index to follow up in days
}
