% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_functions.R
\name{derive_followup_time}
\alias{derive_followup_time}
\title{Calculate followup_time, the number of days from index to follow up in days}
\usage{
derive_followup_time(
  cohort,
  index_date = "index_date",
  followup_date = "followup_date",
  followup_time_rename = "followup_days"
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe.}

\item{index_date}{character. The name of a column in cohort indicating index dates.}

\item{followup_date}{character. The name of a column in cohort indicating follow up dates.}

\item{followup_time_rename}{character. (Optional) Name of the resulting follow up days column.}
}
\value{
dataframe.
}
\description{
Calculate followup_time, the number of days from index to follow up in days
}
