% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_cohort.R
\docType{data}
\name{ez_cohort}
\alias{ez_cohort}
\title{A simulated dataset containing the survival data of 100 patients}
\format{
A data frame with 1626 rows and 2 variables:
\describe{
  \item{patientid}{character. Unique patientid}
  \item{condition}{character. The condition a patient was in.}
  \item{lotstartdate}{date. Date at which first line of therapy starts. used
  as an indexdate}
  \item{lastvisitdate}{date. Last known patient contactdate}
  \item{dateofdeath}{date. Date of death if known.}
}
}
\usage{
ez_cohort
}
\description{
A simulated dataset containing the survival data of 100 patients
}
\keyword{datasets}
