% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_event_name}
\alias{get_event_name}
\title{Get the event name from an iwillsurvive object}
\usage{
get_event_name(object)
}
\arguments{
\item{object}{}
}
\value{
character.
}
\description{
Get the event name from an iwillsurvive object
}
