% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_index_name}
\alias{get_index_name}
\title{Get the index name from an iwillsurvive object}
\usage{
get_index_name(object)
}
\arguments{
\item{object}{}
}
\value{
character.
}
\description{
Get the index name from an iwillsurvive object
}
