% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival.R
\name{iwillsurvive}
\alias{iwillsurvive}
\title{Create an iwillsurvive object}
\usage{
iwillsurvive(
  cohort,
  followup_time = "followup_months",
  event_status = "event_status",
  patient_id = "patientid",
  terms = NULL,
  type = "right",
  fit = NULL,
  event_title = NULL,
  index_title = NULL,
  title = NULL,
  followup_time_units = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe}

\item{followup_time}{character.}

\item{event_status}{character.}

\item{patient_id}{character.}

\item{terms}{character.}

\item{type}{character. See ?survival::survfit.formula}

\item{fit}{survfit. An optional survfit object created from
survival:survfit(). If specified, no additional model(s) will be fit.}

\item{event_title}{character.}

\item{index_title}{character.}

\item{title}{character.}

\item{followup_time_units}{character. Units used in followup time}

\item{verbose}{logical. If TRUE, return messages}
}
\value{
survfit. See ?survival::survfit.object to learn more
}
\description{
Create an iwillsurvive object
}
\examples{

cohort <- data.frame(
  patientid = 1:20,
  followup_months = c(
    6.1, 15.4, 22, 24.6, 25.6, 26.1, 28.7, 46.9, 54.5, 55, 62.2,
    65.5, 88.1, 108.5, 116, 119.1, 119.6, 169.1, 317.8, 381.7
  ),
  event_status = c(
    FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE,
    TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE
  ),
  group = c(
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b",
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b"
  )
)

# Ignoring group term

cohort_fit <- iwillsurvive(cohort)
cohort_fit

# Now including some covariates

cohort_fit <- iwillsurvive(cohort,
                           terms = "group")

cohort_fit
}
