% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_iwillsurvive.R
\name{plot.iwillsurvive}
\alias{plot.iwillsurvive}
\title{Plot an iwillsurvive x.}
\usage{
\method{plot}{iwillsurvive}(
  x = NULL,
  ggtheme = ggplot2::theme_bw(),
  palette = "Set1",
  simple = FALSE,
  title = NULL,
  subtitle = NULL,
  censor_pch = "|",
  censor_size = 3,
  add_gridlines = TRUE,
  add_confidence = TRUE,
  add_censor_ticks = TRUE,
  add_labels = TRUE,
  add_median = TRUE,
  add_median_delta = TRUE,
  anchor_arrow = FALSE,
  legend_position = "inside",
  legend_anchor_y = 0.5,
  legend_nudge_y = NULL,
  legend_position_x = NULL,
  xlim = NULL,
  x_breaks = NULL,
  label_size = 3,
  label_color = grDevices::gray(0),
  median_flag_nudge_y = 0.1,
  median_flag_thickness = 0.7,
  risk_table = TRUE,
  risk_table_title = NULL,
  risk_size = 3.5,
  risk_label_size = 1.25,
  index_title = NULL,
  event_title = NULL,
  median_flag_size = 4,
  event_nudge_y = 0.15,
  panel_heights = c(3, 1),
  ...
)
}
\arguments{
\item{x}{iwillsurvive. An iwillsurvive object created from \code{iwillsurvive()}}

\item{ggtheme}{theme. A ggplot2 theme}

\item{palette}{character. The name of a paleete. See ?ggplot2::scale_colour_brewer for examples}

\item{simple}{logical. If TRUE, only plot the Kaplan-Meier estimate}

\item{title}{character. Plot title}

\item{subtitle}{character. Plot subtitle}

\item{censor_pch}{character. The point shape for censor ticks. See \code{?points}.}

\item{censor_size}{numeric. Size of censor ticks.}

\item{add_gridlines}{logical. If TRUE, include gridlines}

\item{add_confidence}{logical. If TRUE, include a confidence interval}

\item{add_censor_ticks}{logical. If TRUE, show censoring tick marks}

\item{add_labels}{logical. If TRUE, show verbal labels}

\item{add_median}{logical. If TRUE, show median survival}

\item{add_median_delta}{logical.}

\item{anchor_arrow}{logical. If TRUE, use an arrow in pointing to the anchor}

\item{legend_position}{character. Where should the strata labels be located?
Either 'inside' for inside the plot, or 'top', or 'right'}

\item{legend_anchor_y}{numeric. Y locations of anchors for legends.
Only used if legend_position = "inside"}

\item{legend_nudge_y}{numeric.}

\item{legend_position_x}{numeric. X position of the legend(s)}

\item{xlim}{numeric.}

\item{x_breaks}{numeric. Major breaks for the x-axis}

\item{label_size}{numeric. Size of the labels.}

\item{label_color}{character. Color of labels.}

\item{median_flag_nudge_y}{numeric. Amount to nudge median label.}

\item{median_flag_thickness}{numeric. Thickness of the flag border}

\item{risk_table}{logical. If TRUE, include the risk table}

\item{risk_table_title}{character. Title for the risk table}

\item{risk_size}{numeric. Size of font in risk table.}

\item{risk_label_size}{numeric. Size of labels in risk table}

\item{index_title}{character.}

\item{event_title}{character.}

\item{median_flag_size}{numeric.}

\item{event_nudge_y}{numeric.}

\item{panel_heights}{numeric. Heights of the KM and Risk panels}

\item{...}{currently ignored}
}
\value{
ggplot2
}
\description{
Plot an iwillsurvive x.
}
\examples{
# Set things up by creating an iwillsurvive x

cohort <- cohort_raw \%>\%
  derive_followup_date(
    event_date = "dateofdeath",
    censor_date = "lastvisitdate"
  ) \%>\%
  derive_followup_time(index_date = "lotstartdate") \%>\%
  derive_event_status(event_date = "dateofdeath")

cohort_iws <- iwillsurvive(cohort,
  followup_time = "followup_days",
  terms = "condition",
  event_title = "Death",
  index_title = "LOT1 Start"
)

# Default plot
plot(cohort_iws)

# Simpler version
plot(cohort_iws,
  simple = TRUE,
  risk_table = FALSE,
  add_confidence = FALSE
)

# Customized version
plot(cohort_iws,
  add_confidence = FALSE,
  add_median_delta = FALSE,
  censor_pch = 3,
  censor_size = 5,
  legend_position_x = c(600, 400),
  legend_nudge_y = c(.25, .3),
  median_flag_nudge_y = .15,
  anchor_arrow = TRUE,
  palette = "Dark2",
  title = "My Title",
  subtitle = "My Subttitle",
  risk_table_title = "My Risk Table Title"
)
}
