% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_follow_up_time}
\alias{plot_follow_up_time}
\title{Plot a patient timeline containing}
\usage{
plot_follow_up_time(
  cohort = NULL,
  follow_up_time = "follow_up_time",
  event_status = "event_status",
  follow_up_time_units = NULL,
  event_name = NULL,
  index_name = "index",
  add_median = TRUE,
  point_size = 1,
  line_size = 0.5,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{cohort}{dataframe. A ORPP dataframe containing columns representing time at
risk and event status}

\item{follow_up_time}{character. Name of a column in cohort}

\item{event_status}{character. Name of a column in cohort}

\item{follow_up_time_units}{character. The units that time at risk are calculated in}

\item{event_name}{character. Name of the event such as "death", "next treatment"}

\item{index_name}{character. Name of the index such as "LOT1 Start" or "Metastatic Diagnosis"}

\item{add_median}{logical. If TRUE, add a line showing median.}

\item{point_size}{numeric.}

\item{line_size}{numeric.}

\item{theme}{A ggplot2 theme}
}
\value{
ggplot2
}
\description{
Plot a patient timeline containing
}
\examples{

cohort <- data.frame(
  patientid = 1:20,
  follow_up_time = c(
    6.1, 15.4, 22, 24.6, 25.6, 26.1, 28.7, 46.9, 54.5, 55, 62.2,
    65.5, 88.1, 108.5, 116, 119.1, 119.6, 169.1, 317.8, 381.7
  ),
  event_status = c(
    FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE,
    TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE
  )
)

plot_follow_up_time(cohort,
  follow_up_time_units = "days",
  event_name = "Death",
  index_name = "LOT1 Start"
)
}
