% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_iwillsurvive.R
\name{print.iwillsurvive}
\alias{print.iwillsurvive}
\title{Print method for an iwillsurvive object}
\usage{
\method{print}{iwillsurvive}(x, ...)
}
\arguments{
\item{x}{iwillsurvive. An iwillsurvive object created from \code{iwillsurvive}}

\item{...}{currently not used.}
}
\description{
Print method for an iwillsurvive object
}
