% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ascii_survival.R
\name{print_ascii_survival}
\alias{print_ascii_survival}
\title{Print an ascii survival curve}
\usage{
print_ascii_survival(fit)
}
\arguments{
\item{fit}{fit. A survival object created by survival::survfit()}
}
\description{
Print an ascii survival curve
}
\examples{

fit <- survival::survfit(
 survival::Surv(followup_days, event_status,
                type = "right"
 ) ~ condition,
 data = cohort_survival
)

print_ascii_survival(fit)

}
