% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_follow_up_time.R
\name{derive_event_status}
\alias{derive_event_status}
\title{Calculate event_status, a logical column indicating whether or not an event occured.}
\usage{
derive_event_status(
  cohort = NULL,
  event_date = NULL,
  event_status_rename = "event_status"
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe.}

\item{event_date}{character. The name of a column in cohort indicating event dates.}

\item{event_status_rename}{character. (Optional) Name of the resulting event_status column}
}
\value{
dataframe.
}
\description{
Calculate event_status, a logical column indicating whether or not an event occured.
}
