% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_followup_time.R, R/derive_functions.R
\name{derive_followup_time}
\alias{derive_followup_time}
\title{Calculate followup_time, the number of days from index to follow up in days}
\usage{
derive_followup_time(
  data,
  index_date = "index_date",
  followup_date = "followup_date",
  followup_time_rename = "followup_days"
)

derive_followup_time(
  data,
  index_date = "index_date",
  followup_date = "followup_date",
  followup_time_rename = "followup_days"
)
}
\arguments{
\item{data}{dataframe. A one-row-per-patient dataframe.}

\item{index_date}{character. The name of a column in data indicating index dates.}

\item{followup_date}{character. The name of a column in data indicating follow up dates.}

\item{followup_time_rename}{character. (Optional) Name of the resulting follow up days column.}
}
\value{
The cohort dataframe with new columns followup_days, followup_months,
followup_years.

dataframe.
}
\description{
Calculate followup_time, the number of days from index to follow up in days

Calculate followup_time, the number of days from index to follow up in days
}
\examples{

data <- cohort_raw \%>\%
  derive_followup_date(
    event_date = "dateofdeath",
    censor_date = "lastvisitdate"
  )

# Derive follow up time from index to follow up in days, months and years.

data \%>\%
  derive_followup_time(
    index_date = "lotstartdate",
    followup_date = "followup_date"
  )
}
\references{
https://pumas.nasa.gov/examples/index.php?id=46
}
