% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_followup_time.R
\name{derive_followup_time}
\alias{derive_followup_time}
\title{Calculate followup_time, the number of days from index to follow up in days}
\usage{
derive_followup_time(
  cohort,
  index_date = "index_date",
  followup_date = "followup_date",
  days_in_month = 30.44,
  days_in_year = 365.24
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe.}

\item{index_date}{character. The name of a column in cohort indicating index dates.}

\item{followup_date}{character. The name of a column in cohort indicating follow up dates.}

\item{days_in_month}{numeric. Average number of days in a month. Defaults to
30.44.}

\item{days_in_year}{numeric. Average number of days in a year. Defaults to
365.24}
}
\value{
The cohort dataframe with new columns followup_days, followup_months,
followup_years.
}
\description{
Calculate followup_time, the number of days from index to follow up in days
}
\examples{

cohort <- cohort_raw \%>\%
  derive_followup_date(
    event_date = "dateofdeath",
    censor_date = "lastvisitdate"
  )

# Derive follow up time from index to follow upin days, months and years.

cohort \%>\%
  derive_followup_time(
    index_date = "lotstartdate",
    followup_date = "followup_date"
  )
}
\references{
https://pumas.nasa.gov/examples/index.php?id=46
}
