% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ez_cohort.R
\docType{data}
\name{ez_cohort}
\alias{ez_cohort}
\title{A dataset containing the survival data of 100 simulated patients}
\format{
A data frame with 1626 rows and 2 variables:
\describe{
  \item{PatientID}{character. Unique patientid}
  \item{VisitDate}{date. The date of a visit}
}
}
\usage{
ez_cohort
}
\description{
A dataset containing the survival data of 100 simulated patients
}
\keyword{datasets}
