% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival.R
\name{fit_survival}
\alias{fit_survival}
\title{Fit a survival model with a easy to use interface}
\usage{
fit_survival(
  cohort,
  followup_time = "followup_time",
  event_status = "event_status",
  patient_id = "patientid",
  terms = NULL,
  type = "right",
  event_name = NULL,
  index_name = NULL,
  title = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe}

\item{followup_time}{character.}

\item{event_status}{character.}

\item{terms}{character.}

\item{type}{character. See ?survival::survfit.formula}

\item{verbose}{logical. If TRUE, return messages}
}
\value{
survfit. See ?survival::survfit.object to learn more
}
\description{
This is a wrapper around survival::survfit() with a cleaner interface.
}
\examples{

cohort <- data.frame(
  patientid = 1:20,
  followup_time = c(
    6.1, 15.4, 22, 24.6, 25.6, 26.1, 28.7, 46.9, 54.5, 55, 62.2,
    65.5, 88.1, 108.5, 116, 119.1, 119.6, 169.1, 317.8, 381.7
  ),
  event_status = c(
    FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE,
    TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE
  ),
  group = c(
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b",
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b"
  )
)

# Ignoring group term

cohort_fit <- fit_survival(cohort)
cohort_fit

# Now including some covariates

cohort_fit <- fit_survival(cohort, terms = "group")
cohort_fit
}
