% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_survival.R, R/iwillsurvive.R
\name{iwillsurvive}
\alias{iwillsurvive}
\title{Create an iwillsurvive object}
\usage{
iwillsurvive(
  cohort,
  followup_time = "followup_months",
  event_status = "event_status",
  patient_id = "patientid",
  terms = NULL,
  type = "right",
  fit = NULL,
  event_title = "<event>",
  index_title = "<index>",
  title = NULL,
  followup_time_units = NULL,
  verbose = TRUE
)

iwillsurvive(
  cohort,
  followup_time = "followup_months",
  event_status = "event_status",
  patient_id = "patientid",
  terms = NULL,
  type = "right",
  fit = NULL,
  event_title = "<event>",
  index_title = "<index>",
  title = NULL,
  followup_time_units = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe}

\item{followup_time}{character. The column representing followup time.}

\item{event_status}{character. The column representing event status -- must be logical.}

\item{patient_id}{character. ... uniquep patientid.}

\item{terms}{character. A vector of model terms.}

\item{type}{character. See ?survival::survfit.formula}

\item{fit}{survfit. An optional survfit object created from survival:survfit(). If specified, no additional model(s) will be fit.}

\item{event_title}{character. An optional name of}

\item{index_title}{character.}

\item{title}{character. An optional title of the analysis that is passed to print() and plot() methods.}

\item{followup_time_units}{character. Units used in followup time}

\item{verbose}{logical. If TRUE, return messages}
}
\value{
survfit. See ?survival::survfit.object to learn more

survfit. See ?survival::survfit.object to learn more
}
\description{
Create an iwillsurvive object

Create an iwillsurvive object
}
\examples{

cohort <- data.frame(
  patientid = 1:20,
  followup_months = c(
    6.1, 15.4, 22, 24.6, 25.6, 26.1, 28.7, 46.9, 54.5, 55, 62.2,
    65.5, 88.1, 108.5, 116, 119.1, 119.6, 169.1, 317.8, 381.7
  ),
  event_status = c(
    FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE,
    TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE
  ),
  group = c(
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b",
    "a", "b", "a", "b", "a", "b", "a", "b", "a", "b"
  )
)

# Ignoring group term
cohort_fit <- iwillsurvive(cohort)

# Now including some covariates
cohort_fit <- iwillsurvive(cohort,
  terms = "group"
)

cohort_survival

# No Terms

cohort_fit <- iwillsurvive(cohort_survival,
  followup_time = "followup_days"
)
cohort_fit

# Now including a term

cohort_fit <- iwillsurvive(cohort_survival,
  followup_time = "followup_days",
  terms = "condition"
)
cohort_fit
}
