% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iwillsurvive.R
\name{iwillsurvive}
\alias{iwillsurvive}
\title{Create an iwillsurvive object}
\usage{
iwillsurvive(
  cohort,
  followup_time = "followup_months",
  event_status = "event_status",
  patient_id = "patientid",
  terms = NULL,
  type = "right",
  fit = NULL,
  event_title = "<event>",
  index_title = "<index>",
  title = NULL,
  followup_time_units = NULL,
  verbose = TRUE
)
}
\arguments{
\item{cohort}{dataframe. A one-row-per-patient dataframe}

\item{followup_time}{character. The column representing followup time.}

\item{event_status}{character. The column representing event status -- must be logical.}

\item{patient_id}{character. ... uniquep patientid.}

\item{terms}{character. A vector of model terms.}

\item{type}{character. See ?survival::survfit.formula}

\item{fit}{survfit. An optional survfit object created from survival:survfit(). If specified, no additional model(s) will be fit.}

\item{event_title}{character. An optional name of}

\item{index_title}{character.}

\item{title}{character. An optional title of the analysis that is passed to print() and plot() methods.}

\item{followup_time_units}{character. Units used in followup time}

\item{verbose}{logical. If TRUE, return messages}
}
\value{
survfit. See ?survival::survfit.object to learn more
}
\description{
Create an iwillsurvive object
}
\examples{

cohort_survival

# No Terms

cohort_fit <- iwillsurvive(cohort_survival,
                          followup_time = "followup_days")
cohort_fit

# Now including a term

cohort_fit <- iwillsurvive(cohort_survival,
                           followup_time = "followup_days",
                           terms = "condition")
cohort_fit

}
