% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_followup.R
\name{plot_followup}
\alias{plot_followup}
\title{Plot a patient timeline containing}
\usage{
plot_followup(
  object = NULL,
  add_median = TRUE,
  point_size = 1,
  line_size = 0.5,
  theme = ggplot2::theme_minimal()
)
}
\arguments{
\item{object}{iwillsurvive. An iwillsurvive object}

\item{add_median}{logical. If TRUE, add a line showing median.}

\item{point_size}{numeric.}

\item{line_size}{numeric.}

\item{theme}{A ggplot2 theme}

\item{followup_time_units}{= "days",}
}
\value{
ggplot2
}
\description{
Plot a patient timeline containing
}
\examples{

}
