% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot the results of a survival analysis}
\usage{
plot_survival(
  fit = NULL,
  cohort = NULL,
  event_name = NULL,
  index_name = "index",
  follow_up_time_units = NULL,
  ggtheme = ggplot2::theme_bw(),
  palette = c("#4941D1", "#00B6DA")
)
}
\arguments{
\item{fit}{survfit. A survfit object created from survival::survfit()}

\item{cohort}{dataframe. A one-row-per-patient cohort used in generating fit.}

\item{event_name}{character. Name of the event such as "death", "next treatment"}

\item{index_name}{character. Name of the index such as "LOT1 Start" or "Metastatic Diagnosis"}

\item{follow_up_time_units}{character. The units that time at risk are calculated in}

\item{ggtheme}{theme. A ggplot2 theme}

\item{palette}{character. Colors for the color palette}
}
\value{
ggplot2
}
\description{
Plot the results of a survival analysis
}
\examples{

cohort <- data.frame(
  PatientID = 1:20,
  follow_up_time = c(
    6.1, 15.4, 22, 24.6, 25.6, 26.1, 28.7, 46.9, 54.5, 55, 62.2,
    65.5, 88.1, 108.5, 116, 119.1, 119.6, 169.1, 317.8, 381.7
  ),
  event_status = c(
    FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, FALSE,
    TRUE, TRUE, TRUE, TRUE, FALSE, TRUE, TRUE, TRUE, FALSE, FALSE
  )
)

cohort_fit <- fit_survival(cohort)

plot_survival(
  fit = cohort_fit,
  cohort = cohort,
  event_name = "Death",
  index_name = "Metastatic Diagnosis"
)
}
