% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot the results of a survival analysis}
\usage{
plot_survival(
  object = NULL,
  cohort = NULL,
  followup_time_units = NULL,
  ggtheme = ggplot2::theme_bw(),
  palette = c("#4941D1", "#00B6DA"),
  conf_int = TRUE,
  add_median = TRUE,
  add_median_delta = TRUE,
  risk.table = TRUE,
  index_title = NULL,
  event_title = NULL,
  method = "iwillsurvive",
  event_nudge_y = 0.15
)
}
\arguments{
\item{object}{iwillsurvive. An iwillsurvive object created from fit_survival}

\item{cohort}{dataframe. A one-row-per-patient cohort used in generating fit.}

\item{followup_time_units}{character. The units that time at risk are calculated in}

\item{ggtheme}{theme. A ggplot2 theme}

\item{palette}{character. Colors for the color palette}

\item{conf_int}{logical. See ?survminer::ggsurvplot}

\item{add_median}{logical. If TRUE, show median survival}

\item{add_median_delta}{logical.}

\item{risk.table}{logical. See ?survminer::ggsurvplot}

\item{index_title}{character.}

\item{event_title}{character.}

\item{method}{character. Experimental...}

\item{event_nudge_y}{numeric.}
}
\value{
ggplot2
}
\description{
Plot the results of a survival analysis
}
