% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot the results of a survival analysis}
\usage{
plot_survival(
  object = NULL,
  cohort = NULL,
  followup_time_units = NULL,
  ggtheme = ggplot2::theme_bw(),
  palette = c("#4941D1", "#00B6DA"),
  conf.int = TRUE,
  risk.table = TRUE
)
}
\arguments{
\item{object}{iwillsurvive. An iwillsurvivel object created from fit_survival}

\item{cohort}{dataframe. A one-row-per-patient cohort used in generating fit.}

\item{followup_time_units}{character. The units that time at risk are calculated in}

\item{ggtheme}{theme. A ggplot2 theme}

\item{palette}{character. Colors for the color palette}

\item{conf.int}{logical. See ?survminer::ggsurvplot}

\item{risk.table}{logical. See ?survminer::ggsurvplot}
}
\value{
ggplot2
}
\description{
Plot the results of a survival analysis
}
