% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_survival.R
\name{plot_survival}
\alias{plot_survival}
\title{Plot the results of a survival analysis}
\usage{
plot_survival(
  fit = NULL,
  cohort = NULL,
  event_name = NULL,
  index_name = "index",
  followup_time_units = NULL,
  ggtheme = ggplot2::theme_bw(),
  palette = c("#4941D1", "#00B6DA")
)
}
\arguments{
\item{fit}{survfit. A survfit object created from survival::survfit()}

\item{cohort}{dataframe. A one-row-per-patient cohort used in generating fit.}

\item{event_name}{character. Name of the event such as "death", "next treatment"}

\item{index_name}{character. Name of the index such as "LOT1 Start" or "Metastatic Diagnosis"}

\item{followup_time_units}{character. The units that time at risk are calculated in}

\item{ggtheme}{theme. A ggplot2 theme}

\item{palette}{character. Colors for the color palette}
}
\value{
ggplot2
}
\description{
Plot the results of a survival analysis
}
